Updated the deploy default

+++ resources/views/scripts/apps/laravel/laravel-deploy-default.blade.php
@@ -6,12 +6,15 @@ composer install --no-interaction --prefer-dist --optimize-autoloader
 
 echo "" | sudo -S service php7.2-fpm reload
 
-## Copy the .env.example file
-if [ -f .env.example ]; then
-    cp .env.example .env
+if [ ! -f .env ]; then
+    ## Copy the .env.example file
+    if [ -f .env.example ]; then
+        cp .env.example .env
+    fi
 fi
 
 ## Setup Laravel
 if [ -f artisan ]; then
     php artisan key:generate
+    php artisan config:cache
 fi
